Move CRC32C to common
